Moved where the highlight keyword function is called.
The highlight keyword call was moved from the init template frame to the init of the coverletterwindow.  This was because the coverletter window controls the flow of the program and the highlight keywords needs to be called after the rest of the program has been initialized.  Mostly because the cover letterwindow decides when to pack the UI for the template frame and if the call to the highlight keyword function is before the packing then it does nothing.  So it was better to call it from the cover letter window when the program was ready.